fix: relation in db

+++ backend/src/main/java/com/GDU/backend/models/Category.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.Set;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -15,11 +17,12 @@ import lombok.NoArgsConstructor;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Byte id;
 
     @Column(name = "category_name", nullable = false, length = 50)
     private String categoryName;
 
     @Column(name = "category_slug", nullable = false, length = 100)
     private String categorySlug;
+
 }
